chore: rename scripts
