+++ server/src/jobs/importInserJeunesEtablissementsData.js
@@ -3,29 +3,26 @@ const logger = require("../common/logger").child({ context: "import" });
 const { compose, mergeStreams, writeData, oleoduc, transformData } = require("oleoduc");
 const { readCSV } = require("../common/utils/streamUtils");
 const { dbCollection } = require("../common/mongodb");
+const { findRegionByName } = require("../common/regions");
+const { pick } = require("lodash");
 const { omitNil } = require("../common/utils/objectUtils");
 const { asInteger } = require("../common/utils/stringUtils");
 
-const defaultStream = async () => {
-  return mergeStreams([
-    compose(
-      readCSV(await getFromStorage("depp-2022-france-voie-pro-sco-educ-nat-id-mefstat11-2020-maj2-112295-2020.csv")),
-      transformData((line) => {
-        return { ...line, type: "pro", millesime: "2020-2019" };
-      })
-    ),
-    compose(
-      readCSV(await getFromStorage("depp-2022-france-apprentissage-id-formation-2020-maj2-112292-2020.csv")),
-      transformData((line) => {
-        return { ...line, type: "apprentissage", millesime: "2020-2019" };
-      })
-    ),
-  ]);
-};
+async function defaultStream() {
+  return compose(
+    mergeStreams([
+      readCSV(await getFromStorage("depp-2022-etablissement-voie-pro-sco-2020-2019-maj2-112307.csv")),
+      readCSV(await getFromStorage("depp-2022-etablissement-apprentissage-2020-2019-maj2-112304.csv")),
+    ]),
+    transformData((line) => {
+      return { ...line, millesime: "2020-2019" };
+    })
+  );
+}
 
-const importNationalData = async (options = {}) => {
-  const stats = { total: 0, created: 0, updated: 0, failed: 0 };
-  const stream = options.input ? readCSV(options.input) : await defaultStream();
+async function importInserJeunesEtablissementsData(options = {}) {
+  let stats = { total: 0, created: 0, updated: 0, failed: 0 };
+  let stream = options.input ? readCSV(options.input) : await defaultStream();
 
   await oleoduc(
     stream,
@@ -33,34 +30,34 @@ const importNationalData = async (options = {}) => {
       try {
         stats.total++;
 
-        const codeFormation = line["Code Formation"];
-        const millesime = line.millesime || options.millesime;
-        const selector = {
+        let codeFormation = line["Code formation apprentissage"] || line["Code formation Mefstat11"];
+        let millesime = line.millesime || options.millesime;
+        let region = findRegionByName(line["Région"]);
+        let selector = {
+          uai_de_etablissement: line["n°UAI de l'établissement"],
           code_formation: codeFormation,
           millesime,
         };
 
-        const res = await dbCollection("inserJeunesNational").updateOne(
+        let res = await dbCollection("inserJeunesEtablissements").updateOne(
           selector,
           {
             $setOnInsert: {
               millesime,
-              type: line.type ?? options.type,
+              type: line["Code formation apprentissage"] ? "apprentissage" : "pro",
+              uai_de_etablissement: line["n°UAI de l'établissement"],
               code_formation: codeFormation,
             },
             $set: omitNil({
+              libelle_de_etablissement: line["Libellé de l'établissement"],
+              region: pick(region, ["code", "nom"]),
               type_de_diplome: line["Type de diplôme"],
               libelle_de_la_formation: line["Libellé de la formation"],
               duree_de_formation: asInteger(line["Durée de formation (en année)"]),
               diplome_renove_ou_nouveau: line["Diplôme renové ou nouveau"],
               taux_de_poursuite_etudes: asInteger(line["Taux de poursuite d'études"]),
-              nb_en_poursuite_etudes: asInteger(line["nb en poursuite d'études"]),
-              nb_en_annee_terminale: asInteger(line["nb en année terminale"]),
               taux_emploi_6_mois_apres_la_sortie: asInteger(line["Taux d'emploi 6 mois après la sortie"]),
-              nb_en_emploi_6_mois_apres_la_sortie: asInteger(line["nb en emploi 6 mois après la sortie"]),
-              nb_sortants_6_mois_apres_la_sortie: asInteger(line["nb de sortants 6 mois après la sortie"]),
               taux_emploi_12_mois_apres_la_sortie: asInteger(line["Taux d'emploi 12 mois après la sortie"]),
-              nb_en_emploi_12_mois_apres_la_sortie: asInteger(line["nb en emploi 12 mois après la sortie"]),
             }),
           },
           { upsert: true }
@@ -84,6 +81,6 @@ const importNationalData = async (options = {}) => {
   );
 
   return stats;
-};
+}
 
-module.exports = importNationalData;
+module.exports = importInserJeunesEtablissementsData;
