+++ server/tests/jobs/importInserJeunesEtablissementsData-test.js
@@ -1,20 +1,20 @@
 const assert = require("assert");
-const importInsertJeunes = require("../../src/jobs/importInsertJeunes");
+const importInserJeunesEtablissementsData = require("../../src/jobs/importInserJeunesEtablissementsData");
 const { createStream } = require("../utils/testUtils");
 const { dbCollection } = require("../../src/common/mongodb");
 
-describe("importInsertJeunes", () => {
-  it("Vérifie qu'on peut ajouter les données du fichier insertJeunes (apprentissage)", async () => {
+describe("importInserJeunesEtablissementsData", () => {
+  it("Vérifie qu'on peut ajouter les données du fichier inserJeunes (apprentissage)", async () => {
     let input =
       createStream(`n°UAI de l'établissement;Libellé de l'établissement;Région;Code formation apprentissage;Type de diplôme;Libellé de la formation;Durée de formation (en année);Diplôme renové ou nouveau;Taux de poursuite d'études;Taux d'emploi 6 mois après la sortie;Taux d'emploi 12 mois après la sortie
 0751234J;Centre de Formation;AUVERGNE-RHONE-ALPES;1022105;MC5;cuisinier en desserts de restaurant;1;non;42;50;58
 `);
 
-    let stats = await importInsertJeunes({ input, millesime: "2022-2021" });
+    let stats = await importInserJeunesEtablissementsData({ input, millesime: "2022-2021" });
 
     assert.deepStrictEqual(stats, { created: 1, failed: 0, total: 1, updated: 0 });
 
-    let docs = await dbCollection("insertJeunes")
+    let docs = await dbCollection("inserJeunesEtablissements")
       .find({}, { projection: { _id: 0 } })
       .toArray();
     assert.strictEqual(docs.length, 1);
@@ -38,17 +38,17 @@ describe("importInsertJeunes", () => {
     });
   });
 
-  it("Vérifie qu'on peut ajouter les données du fichier insertJeunes (pro)", async () => {
+  it("Vérifie qu'on peut ajouter les données du fichier inserJeunes (pro)", async () => {
     let input =
       createStream(`n°UAI de l'établissement;Libellé de l'établissement;Région;Code formation Mefstat11;Type de diplôme;Libellé de la formation;Durée de formation (en année);Diplôme renové ou nouveau;Taux de poursuite d'études;Taux d'emploi 6 mois après la sortie;Taux d'emploi 12 mois après la sortie
 0751234J;Centre de Formation;ILE-DE-FRANCE;23830025007;BAC PRO;maintenance des equipements industriels;3;non;42;50;
 `);
 
-    let stats = await importInsertJeunes({ input, millesime: "2022-2021" });
+    let stats = await importInserJeunesEtablissementsData({ input, millesime: "2022-2021" });
 
     assert.deepStrictEqual(stats, { created: 1, failed: 0, total: 1, updated: 0 });
 
-    let docs = await dbCollection("insertJeunes")
+    let docs = await dbCollection("inserJeunesEtablissements")
       .find({}, { projection: { _id: 0 } })
       .toArray();
     assert.strictEqual(docs.length, 1);
@@ -72,7 +72,7 @@ describe("importInsertJeunes", () => {
   });
 
   it("Vérifie qu'on peut mettre à jour des données", async () => {
-    await importInsertJeunes({
+    await importInserJeunesEtablissementsData({
       millesime: "2022-2021",
       input:
         createStream(`n°UAI de l'établissement;Libellé de l'établissement;Région;Code formation Mefstat11;Type de diplôme;Libellé de la formation;Durée de formation (en année);Diplôme renové ou nouveau;Taux de poursuite d'études;Taux d'emploi 6 mois après la sortie;Taux d'emploi 12 mois après la sortie
@@ -80,7 +80,7 @@ describe("importInsertJeunes", () => {
 `),
     });
 
-    let stats = await importInsertJeunes({
+    let stats = await importInserJeunesEtablissementsData({
       millesime: "2022-2021",
       input:
         createStream(`n°UAI de l'établissement;Libellé de l'établissement;Région;Code formation Mefstat11;Type de diplôme;Libellé de la formation;Durée de formation (en année);Diplôme renové ou nouveau;Taux de poursuite d'études;Taux d'emploi 6 mois après la sortie;Taux d'emploi 12 mois après la sortie
@@ -89,7 +89,7 @@ describe("importInsertJeunes", () => {
     });
 
     assert.deepStrictEqual(stats, { created: 0, failed: 0, total: 1, updated: 1 });
-    let found = await dbCollection("insertJeunes").findOne();
+    let found = await dbCollection("inserJeunesEtablissements").findOne();
     assert.strictEqual(found.libelle_de_etablissement, "Nouveau Centre de Formation");
   });
 
@@ -99,7 +99,7 @@ describe("importInsertJeunes", () => {
 0751234J;Centre de Formation;INVALID;23830025007;BAC PRO;maintenance des equipements industriels;3;non;42;50;58
 `);
 
-    let stats = await importInsertJeunes({ input, millesime: "2022-2021" });
+    let stats = await importInserJeunesEtablissementsData({ input, millesime: "2022-2021" });
 
     assert.deepStrictEqual(stats, { created: 0, failed: 1, total: 1, updated: 0 });
   });
